move BasicEditor to subfolder
